CCMSPUI-488: Added additional controller tests for /amendments client details endpoints
