+++ src/main/java/uk/gov/laa/ccms/caab/advice/ActiveCaseModelAdvice.java
@@ -10,17 +10,17 @@ import uk.gov.laa.ccms.caab.controller.AssessmentController;
 import uk.gov.laa.ccms.caab.controller.application.section.ApplicationSectionsController;
 import uk.gov.laa.ccms.caab.controller.application.section.ApplicationSubmissionController;
 import uk.gov.laa.ccms.caab.controller.application.section.ApplicationTypeSectionController;
-import uk.gov.laa.ccms.caab.controller.application.section.EditClientAddressDetailsController;
-import uk.gov.laa.ccms.caab.controller.application.section.EditClientAddressDetailsSearchController;
-import uk.gov.laa.ccms.caab.controller.application.section.EditClientBasicDetailsController;
-import uk.gov.laa.ccms.caab.controller.application.section.EditClientContactDetailsController;
-import uk.gov.laa.ccms.caab.controller.application.section.EditClientDeceasedDetailsController;
-import uk.gov.laa.ccms.caab.controller.application.section.EditClientEqualOpportunitiesMonitoringDetailsController;
 import uk.gov.laa.ccms.caab.controller.application.section.EditGeneralDetailsSectionController;
 import uk.gov.laa.ccms.caab.controller.application.section.EditProceedingsAndCostsSectionController;
 import uk.gov.laa.ccms.caab.controller.application.section.EvidenceSectionController;
 import uk.gov.laa.ccms.caab.controller.application.section.OpponentsSectionController;
 import uk.gov.laa.ccms.caab.controller.application.section.ProviderDetailsSectionController;
+import uk.gov.laa.ccms.caab.controller.client.EditClientAddressDetailsController;
+import uk.gov.laa.ccms.caab.controller.client.EditClientAddressDetailsSearchController;
+import uk.gov.laa.ccms.caab.controller.client.EditClientBasicDetailsController;
+import uk.gov.laa.ccms.caab.controller.client.EditClientContactDetailsController;
+import uk.gov.laa.ccms.caab.controller.client.EditClientDeceasedDetailsController;
+import uk.gov.laa.ccms.caab.controller.client.EditClientEqualOpportunitiesMonitoringDetailsController;
 import uk.gov.laa.ccms.caab.controller.client.EditClientSummaryController;
 import uk.gov.laa.ccms.caab.controller.submission.CaseSubmissionController;
