+++ src/test/java/uk/gov/laa/ccms/caab/controller/client/CancelClientRegistrationControllerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.laa.ccms.caab.controller.application.client;
+package uk.gov.laa.ccms.caab.controller.client;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
@@ -13,7 +13,6 @@ import org.mockito.InjectMocks;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import uk.gov.laa.ccms.caab.controller.client.CancelClientRegistrationController;
 
 @ExtendWith(MockitoExtension.class)
 public class CancelClientRegistrationControllerTest {
